// src/sdk/database/orm/orm_seo_setting.ts
export interface SeoSettingModel {
  id: string;
  data_creator: string;
  data_updater: string;
  create_time: string;
  update_time: string;
  setting_key: string;
  site_title: string;
  meta_description: string;
  og_title: string;
  og_description: string;
  og_image_url: string;
  canonical_url: string;
  twitter_handle?: string | null;
}